fix: extract context-builder.ts to avoid GraphAI dependency in context entry point
Move pure context-building functions and DEFAULT_INTERACTIVE_SYSTEM_PROMPT
to a new core/ai/context-builder.ts module with no GraphAI imports.
Both llm.ts and context.ts now import from context-builder.ts, ensuring
that loading mulmocast-preprocessor/context does not pull in GraphAI.
