+++ packages/mulmocast-preprocessor/src/context.ts
@@ -1,3 +1,8 @@
 // Lightweight entry point for context-building utilities (no GraphAI dependency)
-export { buildScriptContent, buildBeatContent, buildScriptMetaContent, scriptToViewerData } from "./core/ai/llm.js";
-export { DEFAULT_INTERACTIVE_SYSTEM_PROMPT } from "./core/ai/command/query/prompts.js";
+export {
+  buildScriptContent,
+  buildBeatContent,
+  buildScriptMetaContent,
+  scriptToViewerData,
+  DEFAULT_INTERACTIVE_SYSTEM_PROMPT,
+} from "./core/ai/context-builder.js";
